crypto: update const type (#221)

+++ src/crypto.v
@@ -4,7 +4,7 @@ import rand
 import crypto.rand as crypto_rand
 import encoding.base64
 
-pub const max_safe_unsigned_integer = 4_294_967_295
+pub const max_safe_unsigned_integer = u32(4_294_967_295)
 
 pub fn set_rand_crypto_safe_seed() {
 	first_seed := generate_crypto_safe_int_u32()
